Completed

+++ public/scripts/script.js
@@ -4,6 +4,7 @@ function deletePost(id) {
       document.getElementById(id).remove();
     });
   });
+  location.reload();
 }
 
 const delButtons = document.getElementsByTagName("button");
@@ -11,6 +12,9 @@ const delButtons = document.getElementsByTagName("button");
 for (const button of delButtons) {
   const postID = button.parentElement.parentElement.id;
   button.addEventListener("click", () => {
-    deletePost(postID);
+    if (window.confirm("Are you sure?")) deletePost(postID);
+    else {
+      alert("Operation cancelled");
+    }
   });
 }
